src/spotify/requirements-txt.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# AWS SDK
boto3==1.34.14
aioboto3==12.3.0

# Async HTTP
aiohttp==3.9.1
httpx==0.25.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Spotify Integration
spotipy==2.23.0  # Optional - for reference implementation
requests==2.31.0
urllib3==2.1.0

# Email
aiosmtplib==3.0.1

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Rate limiting
slowapi==0.1.9

# API documentation
pydantic-openapi-schema==1.5.1
